Fix start time calculation

+++ mavlink_control.cpp
@@ -159,7 +159,7 @@ int main(int argc, char** argv) {
     printf("\n\nHit Enter to start:\n\n");
     char key = getchar();
     uint64_t user_start_time_us = get_time_usec();
-    uint64_t user_relative_start_time_sec = (float) (user_start_time_us - gimbal.get_time_of_first_heartbeat_us()) / 1000000.0f;
+    float user_relative_start_time_sec = (float) (user_start_time_us - gimbal.get_time_of_first_heartbeat_us()) / 1000000.0f;
     printf(
       "User start time = %ld us. User start time relative to first heartbeat = %.3f s\n", 
       user_start_time_us,
